requirements.txt
kallisti-core~=1.3.0
django~=2.2.0
djangorestframework~=3.11.0
drf-yasg~=1.20.0
huey~=2.3.0
supervisor~=4.2.0
